Adding Torch block
